Crate dyn_fn

Utilities to work with dyn [Async]Fn*.

Think about Box<dyn Fn>, but with parametrizable storage (Box, Arc, etc.), and supporting asynchronous closure!

Raw storage notably doesn’t require allocation, making it ideally suited for memory-constrained environments.

This crate relies on higher_kinded_types, reexported as hkt, to support generic lifetime in function parameters and/or return type. However, because of a current limitation of the compiler, every closure requires a second PhantomData parameter to carry the lifetime of the argument for the return type. This ergonomic issue doesn’t impact performance.

Examples

No allocation dynamic callback

#![no_std]
use dyn_fn::{LocalDynFn, hkt, storage};

type Callback<'a> = LocalDynFn<'a, hkt::ForRef<str>, hkt::ForFixed<()>, storage::Raw<32>>;
let mut callbacks = heapless::Vec::<Callback, 4>::new();
callbacks.push(Callback::new(|s, _| {
    defmt::debug!("callback called with '{}'", s)
}));

let input = "input";
for cb in &callbacks {
    cb.call(input);
}
// logs "callback called with 'input'"

Asynchronous dynamic callback

use std::time::Duration;

use dyn_fn::{LocalDynAsyncFn, hkt, storage};
use futures_util::future::join_all;

type Callback<'a> = LocalDynAsyncFn<'a, hkt::ForFixed<Duration>, hkt::ForFixed<()>>;
let mut callbacks = Vec::<Callback>::new();
callbacks.push(Callback::new(async |timeout, _| {
    tokio::time::sleep(timeout).await
}));

let timeout = Duration::from_millis(1);
join_all(callbacks.iter().map(|cb| cb.call(timeout))).await;

Re-exports

pub use higher_kinded_types as hkt;

Modules

storage

The storages available for the dynamic functions/returned futures.

Structs

DynAsyncFn

A dynamic AsyncFn stored in FnStorage, whose returned future is stored in FutureStorage.

DynAsyncFnMut

A dynamic AsyncFnMut stored in FnStorage, whose returned future is stored in FutureStorage.

DynAsyncFnOnce

A dynamic AsyncFnOnce stored in FnStorage, whose returned future is stored in FutureStorage.

DynFn

A dynamic Fn stored in FnStorage.

DynFnMut

A dynamic FnMut stored in FnStorage.

DynFnOnce

A dynamic FnOnce stored in FnStorage.

LocalDynAsyncFn

DynAsyncFn, but without the Send + Sync requirement.

LocalDynAsyncFnMut

DynAsyncFnMut, but without the Send + Sync requirement.

LocalDynAsyncFnOnce

DynAsyncFnOnce, but without the Send + Sync requirement.

LocalDynFn

DynFn, but without the Send + Sync requirement.

LocalDynFnMut

DynFnMut, but without the Send + Sync requirement.

LocalDynFnOnce

DynFnOnce, but without the Send + Sync requirement.

Traits

AsyncFnMutSend

A Send + Sync AsyncFnMut whose returned future is Send

AsyncFnOnceSend

A Send + Sync AsyncFnOnce whose returned future is Send

AsyncFnSend

A Send + Sync AsyncFn whose returned future is Send