dyn_fn

Module storage

The storages available for the dynamic functions/returned futures.

Structs

Arcalloc

A type-erased Arc.

Boxalloc

A type-erased Box.

Raw

A raw storage, where object is stored in place.

RawOrBox

A Raw storage with Box backup if the object doesn’t fit in.

Rcalloc

A type-erased Rc.

Traits

Storage

A storage that can be used to store dynamic type-erased objects.

StorageMut

A Storage whose mutable access gives mutable access to the stored object.

StorageSend

A storage implementing Send + Sync if the stored object implements Send + Sync.

Type Aliases

DefaultFnStoragealloc

Default function storage.

DefaultFutureStorage

Default future storage.