dyn_fn

Struct DynAsyncFnOnce

pub struct DynAsyncFnOnce<'capture, Arg: ForLt + 'static, Ret: ForLt + 'static = ForFixed<()>, FnStorage: StorageMut + StorageSend = DefaultFnStorage, FutureStorage: StorageMut = DefaultFutureStorage>(/* private fields */);

A dynamic AsyncFnOnce stored in FnStorage, whose returned future is stored in FutureStorage.

Using Raw/RawOrBox storage avoids the need to allocate the returned future, which results in better performance — RawOrBox may fall back to Box allocation, but if the returned future is big enough to require it, the allocation cost may be negligible compared to polling it.

DynAsyncFnOnce can also be initialized with a synchronous function, in which case call_try_sync offers a lot better performance than call.

Implementations

impl<'capture, Arg: ForLt + 'static, Ret: ForLt + 'static, FnStorage: StorageMut + StorageSend, FutureStorage: StorageMut> DynAsyncFnOnce<'capture, Arg, Ret, FnStorage, FutureStorage>

pub fn is_sync(&self) -> bool

Returns whether the underlying function is synchronous.

pub async fn call<'a>(self, arg: Arg::Of<'a>) -> Ret::Of<'a>

Calls the underlying function.

pub fn call_sync(self, arg: Arg::Of<'_>) -> Option<Ret::Of<'_>>

Calls the underlying function if is synchronous.

pub async fn call_try_sync<'a>(self, arg: Arg::Of<'a>) -> Ret::Of<'a>

Tries calling the underlying function as synchronous, falling back to asynchronous call.

This is equivalent to

if self.is_sync() {
    self.call_sync(arg).unwrap()
} else {
    self.call(arg).await
}

impl<'capture, Arg: ForLt, Ret: ForLt, FnStorage: StorageMut + StorageSend, FutureStorage: StorageMut> DynAsyncFnOnce<'capture, Arg, Ret, FnStorage, FutureStorage>

pub fn new<F: AsyncFnOnceSend<'capture, Arg, Ret>>(f: F) -> Self

Construct a new DynAsyncFnOnce.

impl<'capture, Arg: ForLt, Ret: ForLt, const SIZE: usize, const ALIGN: usize, FutureStorage: StorageMut> DynAsyncFnOnce<'capture, Arg, Ret, Raw<SIZE, ALIGN>, FutureStorage>
where Align<ALIGN>: Alignment,

pub const fn new_box<F: AsyncFnOnceSend<'capture, Arg, Ret>>(f: Box<F>) -> Self

Available on crate feature alloc only.

Construct a new DynAsyncFnOnce.

impl<'capture, Arg: ForLt, Ret: ForLt, FutureStorage: StorageMut> DynAsyncFnOnce<'capture, Arg, Ret, Box, FutureStorage>

pub fn new_raw<F: AsyncFnOnceSend<'capture, Arg, Ret>>(f: Box<F>) -> Self

Available on crate feature alloc only.

Construct a new DynAsyncFnOnce from a boxed function.

impl<'capture, Arg: ForLt, Ret: ForLt, const SIZE: usize, const ALIGN: usize, FutureStorage: StorageMut> DynAsyncFnOnce<'capture, Arg, Ret, RawOrBox<SIZE, ALIGN>, FutureStorage>
where Align<ALIGN>: Alignment,

pub fn new_box<F: AsyncFnOnceSend<'capture, Arg, Ret>>(f: F) -> Self

Available on crate feature alloc only.

Construct a new DynAsyncFnOnce.

pub fn new_raw<F: AsyncFnOnceSend<'capture, Arg, Ret>>(f: Box<F>) -> Self

Available on crate feature alloc only.

Construct a new DynAsyncFnOnce from a boxed function.

impl<'capture, Arg: ForLt, Ret: ForLt, FnStorage: StorageMut + StorageSend, FutureStorage: StorageMut> DynAsyncFnOnce<'capture, Arg, Ret, FnStorage, FutureStorage>

pub fn new_sync<F: for<'a> FnOnce(Arg::Of<'a>, PhantomData<&'a ()>) -> Ret::Of<'a> + Send + Sync + 'capture>( f: F, ) -> Self

Construct a new DynAsyncFnOnce from as synchronous function.

impl<'capture, Arg: ForLt, Ret: ForLt, const SIZE: usize, const ALIGN: usize, FutureStorage: StorageMut> DynAsyncFnOnce<'capture, Arg, Ret, Raw<SIZE, ALIGN>, FutureStorage>
where Align<ALIGN>: Alignment,

pub const fn new_sync_box<F: for<'a> FnOnce(Arg::Of<'a>, PhantomData<&'a ()>) -> Ret::Of<'a> + Send + Sync + 'capture>( f: Box<F>, ) -> Self

Available on crate feature alloc only.

Construct a new DynAsyncFnOnce from as synchronous function.

impl<'capture, Arg: ForLt, Ret: ForLt, FutureStorage: StorageMut> DynAsyncFnOnce<'capture, Arg, Ret, Box, FutureStorage>

pub fn new_sync_raw<F: for<'a> FnOnce(Arg::Of<'a>, PhantomData<&'a ()>) -> Ret::Of<'a> + Send + Sync + 'capture>( f: Box<F>, ) -> Self

Available on crate feature alloc only.

Construct a new DynAsyncFnOnce from a synchronous boxed function.

impl<'capture, Arg: ForLt, Ret: ForLt, const SIZE: usize, const ALIGN: usize, FutureStorage: StorageMut> DynAsyncFnOnce<'capture, Arg, Ret, RawOrBox<SIZE, ALIGN>, FutureStorage>
where Align<ALIGN>: Alignment,

pub fn new_sync_box<F: for<'a> FnOnce(Arg::Of<'a>, PhantomData<&'a ()>) -> Ret::Of<'a> + Send + Sync + 'capture>( f: F, ) -> Self

Available on crate feature alloc only.

Construct a new DynAsyncFnOnce from as synchronous function.

pub fn new_sync_raw<F: for<'a> FnOnce(Arg::Of<'a>, PhantomData<&'a ()>) -> Ret::Of<'a> + Send + Sync + 'capture>( f: Box<F>, ) -> Self

Available on crate feature alloc only.

Construct a new DynAsyncFnOnce from a synchronous boxed function.